@@ -3,23 +3,15 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.9",
-    "@testing-library/react": "^11.2.5",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/jest": "^26.0.21",
-    "@types/node": "^12.20.13",
-    "@types/react": "^17.0.5",
-    "@types/react-dom": "^17.0.5",
-    "@types/react-redux": "^7.1.16",
     "@types/redux-saga": "^0.10.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-intl": "^5.17.6",
     "react-redux": "^7.2.4",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "4.0.3",
     "reselect": "^4.0.0",
-    "styled-components": "^5.3.0",
+    "styled-components": "^5.3.5",
     "typescript": "^4.2.3",
     "web-vitals": "^1.1.1"
   },
@@ -48,6 +40,14 @@
     ]
   },
   "devDependencies": {
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.1.1",
+    "@testing-library/user-event": "^14.1.1",
+    "@types/jest": "^27.4.1",
+    "@types/node": "^17.0.25",
+    "@types/react": "^18.0.6",
+    "@types/react-dom": "^18.0.2",
+    "@types/react-redux": "^7.1.24",
     "@types/react-router-dom": "^5.1.7",
     "@types/styled-components": "^5.1.9"
   }
