@@ -1,17 +1,12 @@
 /**
  * AppContainer
  */
-import { BrowserRouter } from "react-router-dom";
 
 // import containers
 import AppRoutes from "./AppRoutes";
 
 const AppContainer = () => {
-  return (
-    <BrowserRouter>
-      <AppRoutes />
-    </BrowserRouter>
-  );
+  return <AppRoutes />;
 };
 
 export default AppContainer;
